processor/requirements.txt
pika==1.3.1
requests==2.32.5
python-dotenv==1.2.1
pandas==2.2.3
